+++ src/test/TicTacToken.t.sol
@@ -4,18 +4,18 @@ pragma solidity ^0.8.0;
 import "./utils/TicTacTokenTest.sol";
 
 contract TestTTT is TicTacTokenTest {
-    uint256 internal constant EMPTY = 0;
-    uint256 internal constant X = 1;
-    uint256 internal constant O = 2;
+    uint8 internal constant EMPTY = 0;
+    uint8 internal constant X = 1;
+    uint8 internal constant O = 2;
 
     function test_has_empty_board() public {
-        for (uint256 i = 0; i < 9; i++) {
+        for (uint8 i = 0; i < 9; i++) {
             assertEq(ttt.board(i), 0);
         }
     }
 
     function test_get_board() public {
-        uint256[9] memory expected = [
+        uint8[9] memory expected = [
             EMPTY,
             EMPTY,
             EMPTY,
@@ -26,8 +26,8 @@ contract TestTTT is TicTacTokenTest {
             EMPTY,
             EMPTY
         ];
-        uint256[9] memory actual = ttt.getBoard();
-        for (uint256 i = 0; i < 9; i++) {
+        uint8[9] memory actual = ttt.getBoard();
+        for (uint8 i = 0; i < 9; i++) {
             assertEq(actual[i], expected[i]);
         }
     }
